wpc3

+++ IOT.Core.Repository/CommType/CommTypeRepository.cs
@@ -13,9 +13,17 @@ namespace IOT.Core.Repository.CommType
     {
         public int Delete(string ids)
         {
-            string sql = $"delete from CommType where TId in ({ids})";
-
-            return DapperHelper.Execute(sql);
+            int sss = Convert.ToInt32(DapperHelper.Exescalar($"select count(*) from CommType where ParentId={ids}"));
+            if (sss>0)
+            {
+               string sql = $"delete from CommType where TId in ({ids})"; 
+                return DapperHelper.Execute(sql);
+            }
+            else
+            {
+                return 0;
+            }
+           
         }
 
         public int Insert(Model.CommType Model)
@@ -24,9 +32,13 @@ namespace IOT.Core.Repository.CommType
             return DapperHelper.Execute(sql);
         }
 
-        public List<Model.CommType> Query()
+        public List<Model.CommType> Query(int tid=0)
         {
             string sql = $"select * from CommType";
+            if (tid!=0)
+            {
+                sql += $"  where TId ={tid}";
+            }
             return DapperHelper.GetList<Model.CommType>(sql);
         }
         public List<Model.CommType> UptState(int id)
@@ -39,7 +51,12 @@ namespace IOT.Core.Repository.CommType
             string sql = $"update CommType set TName='{c.TName}',Sort={c.Sort} where TId={c.TId} ";
             return DapperHelper.Execute(sql);
         }
-        
-       
+
+        public List<Model.CommType> Query()
+        {
+            throw new NotImplementedException();
+        }
+
+      
     }
 }
